Bug 1806787: Expose openshift_console_operator_build_info metric
