+++ pkg/console/operator/operator.go
@@ -42,7 +42,6 @@ import (
 	operatorclientv1 "github.com/openshift/client-go/operator/clientset/versioned/typed/operator/v1"
 
 	// operator
-	"github.com/openshift/console-operator/pkg/console/metrics"
 	statushelpers "github.com/openshift/console-operator/pkg/console/status"
 	"github.com/openshift/console-operator/pkg/console/subresource/configmap"
 	"github.com/openshift/console-operator/pkg/console/subresource/deployment"
@@ -70,8 +69,6 @@ type consoleOperator struct {
 	routeClient   routeclientv1.RoutesGetter
 	oauthClient   oauthclientv1.OAuthClientsGetter
 	versionGetter status.VersionGetter
-	// metrics
-	consoleMetrics *metrics.ConsoleMetrics
 	// recorder
 	recorder       events.Recorder
 	resourceSyncer resourcesynccontroller.ResourceSyncer
@@ -98,8 +95,6 @@ func NewConsoleOperator(
 	oauthClients oauthinformersv1.OAuthClientInformer,
 	// openshift managed
 	managedCoreV1 corev1.Interface,
-	// metrics
-	consoleMetrics *metrics.ConsoleMetrics,
 	// event handling
 	versionGetter status.VersionGetter,
 	recorder events.Recorder,
@@ -121,8 +116,6 @@ func NewConsoleOperator(
 		routeClient:   routev1Client,
 		oauthClient:   oauthv1Client,
 		versionGetter: versionGetter,
-		// metrics
-		consoleMetrics: consoleMetrics,
 		// recorder
 		recorder:       recorder,
 		resourceSyncer: resourceSyncer,
