+++ pkg/console/version/version.go
@@ -1,9 +1,8 @@
 package version
 
 import (
+	"github.com/openshift/console-operator/pkg/console/metrics"
 	"k8s.io/apimachinery/pkg/version"
-	"k8s.io/component-base/metrics"
-	"k8s.io/component-base/metrics/legacyregistry"
 )
 
 var (
@@ -34,14 +33,5 @@ func Get() version.Info {
 }
 
 func init() {
-	buildInfo := metrics.NewGaugeVec(
-		&metrics.GaugeOpts{
-			Name: "openshift_console_operator_build_info",
-			Help: "A metric with a constant '1' value labeled by major, minor, git commit & git version from which OpenShift Console Operator was built.",
-		},
-		[]string{"major", "minor", "gitCommit", "gitVersion"},
-	)
-	buildInfo.WithLabelValues(majorFromGit, minorFromGit, commitFromGit, versionFromGit).Set(1)
-
-	legacyregistry.MustRegister(buildInfo)
+	metrics.RegisterVersion(majorFromGit, minorFromGit, commitFromGit, versionFromGit)
 }
